+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0]
+
+### Added
+* Now handles cross-mod compat for FTB Chunks
+  * FTB Chunks 1902.4.0 or later required; all cross-mod compat has been moved from FTB Chunks to FTB XMod Compat
+  * Mod support: KubeJS (6.1), Game Stages (Forge only), Waystone (Forge/Fabric), Common Protection API (Fabric only)
+  
 ## [1.1.2]
 
 ### Fixed